03-googlesheets-exercise/requirements.txt
google-auth ~=2.36
google-api-python-client ~=2.152
